Move environment stuff into it's own submodule
