+++ pytoil/environments/abstract.py
@@ -1,0 +1,6 @@
+"""
+Abstract base class for all Virtual Environments
+
+Author: Tom Fleet
+Created: 07/03/2021
+"""
